+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # Change log
 
-当前分支尚无任何提交
+## Last
+
+### ✨ Features
+
+- init ([27a359f](git@github.com:monako97/expression-language-editor/commit/27a359fb9946a0cb8df086d815c0b6537b69edf1))
+
+## undefined